src/requirements.txt
asgiref==3.8.1
boto3==1.38.11
botocore==1.38.11
certifi==2025.4.26
channels==4.2.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
pipenv==2025.0.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.12.0
setuptools==80.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
whitenoise==6.9.0
